BUG: fix download of GCS manifest based content
Due to the switch from `public-datasets-idc` to `idc-open-data` bucket for
GCS, the logic for detecting provider from s3_url had to be updated.

In the situations where manifest contains a mix of URLs between `idc-open-data`
and other buckets within the same provider, the prior logic would fail since
`idc-open-data` URLs would match those in AWS, but other URLs would not,
and the code would incorrectly conclude that the manifest contains a mix of URLs
across more than one provider.

In the updated code, URLs that refer to `idc-open-data` are ignored, and the
assessment of whether there is more than one endpoint is done for the remaining
URLs. Also, while mapping manifest URLs from AWS to GCP, mapping of `idc-open-data`
is not done, `public-datasets-idc` has been replaced with `idc-open-data` for GCS
in v20, and the former will eventually be deprecated and removed, although for now
it is possible to download from that bucket.